Normalize data
Add simple normalization by dividing by the maximum number of the
training data. Currently I don't care about the minimum number, because
it is above 0 and very close to 0 for the timedifferences. For the PD
measurement values I have to ask why some of them are negative. If they
are not to be considered 0, then the normaliation should probably be
adapted.
To do the normalization, beforehand splitting is necessary, because the
normalization must just be done on the training data and the applied on
the complete data.

Improves on #24