Do scaling after fingerprint build & add cross val
Add cross validation to have more robust accuracy score numbers.
To add cross validation in an easy way perform the normalization after
the scaling has been done.
This might also increase the accuracy itself, because scaling is
actually only required for the features given to the classifier and not
beforehand, so before the fingerprints have been built.