Add 1D DTW NN classifiers working on min length
As in the 2D case I currently convert all the data that they have the
length of the measurement which has the minimum length.
In case of 2D the length is measured in terms of PDs and in the 1D case
in terms of time.

In order to achieve this these changes are necessary:
- Adapt the the time unit notation to allow for a conversion based on
  the set TIME_UNIT
- Correct the test files as they contain different PDs happening on the
  same time. That is not possible (in the experiments).