Fix data splitting to fix seqfinger models
The problem was split could have duration longer than the specified
"duration". This was not intended. It was caused by splitting by the
cumulative sum of the time differences, but actually using the time
differences in a later step. This is fixed by adapting the first time
difference value of each part.