Drop plots as they might cause tkinter crashes
The following exception has already occured a couple of times, Paths
modified for readability:
Exception ignored in: <function Image.__del__ at 0x7fe33a287dc0>
Traceback (most recent call last):
  File "~/.pyenv/versions/3.8.7/lib/python3.8/tkinter/__init__.py",
  line 4020, in __del__
RuntimeError: main thread is not in main loop

Searching for a cause of this error led me to this result:
https://github.com/r9y9/deepvoice3_pytorch/issues/5
It suggest that matplotlib might be causing this. Therefore the plotting
of the scores is removed from this script. It can still easily be done
with the visualize-results scripts.