Allow to only use first n samples
Until now it was not possible to use only the first n samples of each
measurement for the 1D and fingerprint models. Add this functionality by
specifying the repeat option. If the repeat option is not specified a
high constant is used to have the repition high enough. This change
further requires now that the step duration needs to be specified if
splitting has to be done. This was done already before in the config
anyhow, but it now also required by the code.