Fix package installation by restricting a version
The dtaidistance package adds numpy as a dependency with version 2.3.*.
Since the package is built during the installation of my package with
build isolation, numpy with version >= 1.20 is included for it. Numpy
1.20 introduces a breaking change. To prevent the introduction of this
breaking change, restrict the dtaidistance package to be lower than
version 2.3.

See for a further information in a related problem:
https://stackoverflow.com/questions/66060487/valueerror-numpy-ndarray-size-changed-may-indicate-binary-incompatibility-exp